JpaRepository로 변경
