+++ src/main/java/com/sprint/mission/discodeit/repository/BinaryContentRepository.java
@@ -1,14 +1,9 @@
 package com.sprint.mission.discodeit.repository;
 
 import com.sprint.mission.discodeit.entity.BinaryContent;
-import java.util.Optional;
 import java.util.UUID;
+import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BinaryContentRepository {
+public interface BinaryContentRepository extends JpaRepository<BinaryContent, UUID> {
 
-  BinaryContent save(BinaryContent content);
-
-  Optional<BinaryContent> findById(UUID id);
-
-  void delete(UUID id);
 }
