+++ src/main/java/com/sprint/mission/discodeit/repository/ChannelRepository.java
@@ -1,19 +1,9 @@
 package com.sprint.mission.discodeit.repository;
 
 import com.sprint.mission.discodeit.entity.Channel;
-import java.util.List;
-import java.util.Optional;
 import java.util.UUID;
+import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ChannelRepository {
+public interface ChannelRepository extends JpaRepository<Channel, UUID> {
 
-  Channel save(Channel channel);
-
-  Optional<Channel> findById(UUID channelId);
-
-  List<Channel> findAll();
-
-  void updateChannel(Channel channel);
-
-  void deleteChannel(UUID channelId);
 }
