+++ src/main/java/com/sprint/mission/discodeit/repository/MessageRepository.java
@@ -1,22 +1,13 @@
 package com.sprint.mission.discodeit.repository;
 
+import com.sprint.mission.discodeit.entity.Channel;
 import com.sprint.mission.discodeit.entity.Message;
-
 import java.util.List;
-import java.util.Optional;
 import java.util.UUID;
+import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MessageRepository {
+public interface MessageRepository extends JpaRepository<Message, UUID> {
 
-    Message save(Message message);
+  List<Message> findByChannel(Channel channel);
 
-    Optional<Message> findById(UUID messageId);
-
-    List<Message> findByChannelId(UUID channelId);
-
-    void updateMessage(Message message);
-
-    void deleteMessage(UUID messageId);
-
-    void deleteByChannelId(UUID channelId);
 }
