+++ src/main/java/com/sprint/mission/discodeit/repository/ReadStatusRepository.java
@@ -1,22 +1,17 @@
 package com.sprint.mission.discodeit.repository;
 
+import com.sprint.mission.discodeit.entity.Channel;
 import com.sprint.mission.discodeit.entity.ReadStatus;
-
+import com.sprint.mission.discodeit.entity.User;
 import java.util.List;
-import java.util.Optional;
 import java.util.UUID;
+import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ReadStatusRepository {
+public interface ReadStatusRepository extends JpaRepository<ReadStatus, UUID> {
 
-  ReadStatus save(ReadStatus readStatus);
+  List<ReadStatus> findByUser(User user);
 
-  Optional<ReadStatus> findById(UUID id);
+  List<ReadStatus> findByChannel(Channel channel);
 
-  List<ReadStatus> findByUserId(UUID userId);
-
-  List<ReadStatus> findByChannelId(UUID channelId);
-
-  void delete(UUID id);
-
-  void deleteByChannelId(UUID channelId);
+  void deleteByChannel(Channel channel);
 }
