+++ src/main/java/com/sprint/mission/discodeit/repository/UserRepository.java
@@ -1,22 +1,12 @@
 package com.sprint.mission.discodeit.repository;
 
 import com.sprint.mission.discodeit.entity.User;
-
-import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
+import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserRepository {
+public interface UserRepository extends JpaRepository<User, UUID> {
 
-    User save(User user);
+  Optional<User> findByUsername(String username);
 
-    Optional<User> findById(UUID userId);
-
-    Optional<User> findByName(String name);
-
-    List<User> findAll();
-
-    void updateUser(User user);
-
-    void deleteUser(UUID userId);
 }
