+++ src/main/java/com/sprint/mission/discodeit/repository/UserStatusRepository.java
@@ -1,22 +1,9 @@
 package com.sprint.mission.discodeit.repository;
 
 import com.sprint.mission.discodeit.entity.UserStatus;
-
-import java.util.List;
-import java.util.Optional;
 import java.util.UUID;
+import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserStatusRepository {
+public interface UserStatusRepository extends JpaRepository<UserStatus, UUID> {
 
-  UserStatus save(UserStatus userStatus);
-
-  Optional<UserStatus> findById(UUID id);
-
-  Optional<UserStatus> findByUserId(UUID userId);
-
-  List<UserStatus> findAll();
-
-  void delete(UUID id);
-
-  void deleteByUserId(UUID userId);
 }
