--- src/main/java/com/sprint/mission/discodeit/repository/jcf/JCFUserRepository.java
@@ -1,52 +1,0 @@
-package com.sprint.mission.discodeit.repository.jcf;
-
-import com.sprint.mission.discodeit.entity.User;
-import com.sprint.mission.discodeit.repository.UserRepository;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
-import org.springframework.stereotype.Repository;
-
-import java.util.*;
-
-@Repository
-@ConditionalOnProperty(name = "discodeit.repository.type", havingValue = "jcf")
-public class JCFUserRepository implements UserRepository {
-
-  private final Map<UUID, User> data;
-
-  public JCFUserRepository() {
-    data = new HashMap<>(100);
-  }
-
-  @Override
-  public User save(User user) {
-    data.put(user.getId(), user);
-    return user;
-  }
-
-  @Override
-  public Optional<User> findById(UUID userId) {
-    return Optional.ofNullable(data.get(userId));
-  }
-
-  @Override
-  public Optional<User> findByName(String name) {
-    return data.values().stream()
-        .filter(u -> u.getUsername().equals(name))
-        .findFirst();
-  }
-
-  @Override
-  public List<User> findAll() {
-    return new ArrayList<>(data.values());
-  }
-
-  @Override
-  public void updateUser(User user) {
-    data.put(user.getId(), user);
-  }
-
-  @Override
-  public void deleteUser(UUID userId) {
-    data.remove(userId);
-  }
-}
